refactor: remove unused methods and fields
